+++ mojinprince-server/src/model/mod.rs
@@ -1,5 +1,9 @@
 pub mod ai;
+pub mod data;
 pub mod quote;
+pub mod review;
 
 pub use ai::*;
+pub use data::*;
 pub use quote::*;
+pub use review::*;